refactor car owner to own feature page
